Refactor and move files around
