+++ MixItUp.Desktop/Services/DeveloperAPI/Models/UserCurrencyUpdateDeveloperAPIModel.cs
@@ -1,0 +1,11 @@
+﻿using System.Runtime.Serialization;
+
+namespace MixItUp.Desktop.Services.DeveloperAPI.Models
+{
+    [DataContract]
+    public class UserCurrencyUpdateDeveloperAPIModel
+    {
+        [DataMember]
+        public int Amount { get; set; }
+    }
+}
