+++ MixItUp.Desktop/Services/DeveloperAPI/NoCacheHeader.cs
@@ -1,0 +1,16 @@
+﻿using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace MixItUp.Desktop.Services.DeveloperAPI
+{
+    public class NoCacheHeader : DelegatingHandler
+    {
+        async protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            HttpResponseMessage response = await base.SendAsync(request, cancellationToken);
+            response.Headers.Add("Cache-Control", "no-cache");
+            return response;
+        }
+    }
+}
